+++ crazyflie-link/Cargo.toml
@@ -28,6 +28,7 @@ futures = "0.3"
 futures-util = "0.3"
 futures-channel = "0.3"
 rusb = "0.9.4"
+async-trait = "0.1.83"
 
 [dev-dependencies]
 anyhow = "1"
