+++ crazyflie-link/src/lib.rs
@@ -35,6 +35,7 @@
 extern crate bitflags;
 
 mod connection;
+mod crazyradio_connection;
 mod context;
 mod error;
 mod packet;
